WIP: Linux port  - Improvements to the build system so it handles dependencies on Linux  - Various GCC/Clang compiler error/warning fixes
Former-commit-id: 266484ff97b5afbaa71e1ae92abf57dd844dd192

+++ Source/BansheeCore/Animation/BsAnimationManager.cpp
@@ -217,7 +217,7 @@ namespace bs
 				if (soInfo.boneIdx != -1)
 					continue;
 
-				if (soInfo.layerIdx == (UINT32)-1 || soInfo.stateIdx == (UINT32)-1)
+				if (soInfo.layerIdx == -1 || soInfo.stateIdx == -1)
 					continue;
 
 				const AnimationState& state = anim->layers[soInfo.layerIdx].states[soInfo.stateIdx];
